+++ resources/views/pwa/manifest.blade.php
@@ -1,0 +1,47 @@
+<!-- PWA Includes -->
+@php
+    $config = config('pwa.manifest');
+    if (!$config) {
+        throw new exception('PWA Manifest config file not found. Check if the config/pwa.php file exists.');
+    }
+@endphp
+
+<!-- Web Application Manifest -->
+<link rel="manifest" href="{{ route('pwa.manifest') }}">
+<!-- Chrome for Android theme color -->
+<meta name="theme-color" content="{{ $config['theme_color'] }}">
+
+<!-- Add to the home screen for Chrome on Android -->
+<meta name="mobile-web-app-capable" content="{{ $config['display'] == 'standalone' ? 'yes' : 'no' }}">
+<meta name="application-name" content="{{ $config['short_name'] }}">
+<link rel="icon" sizes="{{ data_get(end($config['icons']), 'sizes') }}" href="{{ data_get(end($config['icons']), 'src') }}">
+
+<!-- Add to the home screen for Safari on iOS -->
+<meta name="apple-mobile-web-app-capable" content="{{ $config['display'] == 'standalone' ? 'yes' : 'no' }}">
+<meta name="apple-mobile-web-app-status-bar-style" content="{{  $config['status_bar'] }}">
+<meta name="apple-mobile-web-app-title" content="{{ $config['short_name'] }}">
+<link rel="apple-touch-icon" href="{{ data_get(end($config['icons']), 'src') }}">
+
+<!-- Minimal Splash Screen (optional, use the most common resolution) -->
+<link href="{{ $config['splash']['1125x2436'] }}" media="(device-width: 375px) and (device-height: 812px) and (-webkit-device-pixel-ratio: 3)" rel="apple-touch-startup-image" />
+
+<!-- Tile for Win8 -->
+<meta name="msapplication-TileColor" content="{{ $config['background_color'] }}">
+<meta name="msapplication-TileImage" content="{{ data_get(end($config['icons']), 'src') }}">
+
+<script type="text/javascript">
+    // Initialize the service worker
+    if ('serviceWorker' in navigator) {
+        navigator.serviceWorker.register('/serviceWorker.js', {
+            scope: '.'
+        }).then(function (registration) {
+            // Registration was successful
+            console.log('ServiceWorker registration successful with scope: ', registration.scope);
+            console.log("Online status: " + navigator.onLine);
+        }, function (err) {
+            // registration failed :(
+            console.log('ServiceWorker registration failed: ', err);
+            console.log("Online status: " + navigator.onLine);
+        });
+    }
+</script>