+++ resources/views/pwa/offline.blade.php
@@ -1,0 +1,9 @@
+{{-- <x-pwa-layout> --}}
+
+{{--@section('content')--}}
+
+    <h1>You are not connected to the Internet and now viewing the offline page.</h1>
+
+{{--@endsection--}}
+
+{{-- </x-pwa-layout> --}}