+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 03 Jul 25
+
+### Added
+
+- User migration to remember last tenant
+- Profile editing screens with PWA diagnostics
+- manifest example
+- ServiceWorker.js
+- PWA icons
+- PWA screenshots
+- The actual PWA Diagnostics page
+- All the PWA events
+
 ## Added
 
 - Spatie sortable